+++ esp-hal/CHANGELOG.md
@@ -60,6 +60,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Cache configuration options for ESP32-S2 (#5306)
 - C5: Add PSRAM support (#5317)
 - C61: Add PSRAM support (#5325)
+- MCPWM: Add external sync line and timer sync out support. (#5344)
+- MCPWM: Add capture channel and timer support. (#5344)
+- MCPWM: Added `esp_hal::mcpwm::timer::Timer::set_config` to apply a config to a timer. (#5344)
 
 ### Changed
 
@@ -101,6 +104,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `esp_hal::efuse::chip_revision` has been marked stable (#5287)
 - `esp_hal::efuse::chip_revision` now returns `ChipRevision` (#5287)
 - DMA buffers can now be created using empty buffer slices (#5266)
+- MCPWM: `esp_hal::mcpwm::timer::Timer::start` is now parameterless. Instead, use `esp_hal::mcpwm::timer::Timer::set_config` to apply a config to a timer. (#5344)
+- MCPWM: timers default to the config given by `esp_hal::mcpwm::timer::TimerClockConfig::default`. (#5344)
 
 ### Fixed